File Structure Change
Config file now set as a Json file. Main file broken up: init function seperated in expectation that it may grow. Functions used by main will be found ing mainfunction.go.
