+++ config.json
@@ -1,0 +1,3 @@
+{
+    "database": "Profiles.txt"
+}
